pyproject.toml
[project]
name = "siyuan-note-struct"
version = "0.1.0"
description = "Fetch and display SiYuan note structure"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
]
